// packages/api/src/generated/user/models/concernSummaryResponseCategory.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Ogonggo User API
 * OpenAPI spec version: v1
 */

/**
 * | 값 | code | 설명 |
 * | --- | --- | --- |
 * | `JOB_POSTING` | 1 | 공고 질문 |
 * | `CAREER` | 2 | 직무·커리어 |
 * | `APPLICATION_INTERVIEW` | 3 | 서류·면접 |
 * | `SIDE_EXPERIENCE` | 4 | 사이드·경험 |
 * | `ETC` | 5 | 기타 |
 */
export type ConcernSummaryResponseCategory = typeof ConcernSummaryResponseCategory[keyof typeof ConcernSummaryResponseCategory];


export const ConcernSummaryResponseCategory = {
  JOB_POSTING: 'JOB_POSTING',
  CAREER: 'CAREER',
  APPLICATION_INTERVIEW: 'APPLICATION_INTERVIEW',
  SIDE_EXPERIENCE: 'SIDE_EXPERIENCE',
  ETC: 'ETC',
} as const;